pricing-service/requirements.txt
google-genai>=0.3.0
openai>=1.40.0
flask
